--- src/pages/profile/ProfilePage.js
@@ -1,6 +1,0 @@
-import '/src/components/Atom/profileComponent';
-import '/src/components/Atom/ProfilePhoto';
-import '/src/components/Atom/profileFooter';
-import '/src/components/Molecule/ProfileIconballons/ProfileIconBalloon';
-import '/src/components/Molecule/ProfileTemperature/ProfileTemperature';
-import '/src/components/Molecule/ProfileIconballons/ProfileChatBallon';
