+++ src/pages/profile/index.html
@@ -8,14 +8,15 @@
     <link rel="icon" href="/src/assets/favicon.svg" />
     <link rel="stylesheet" href="/src/style/variable.css" />
     <link rel="stylesheet" href="/src/style/reset.css" />
-    <script type="module" src="/src/pages/profile/ProfilePage.js"></script>
+    <script type="module" src="./index.js"></script>
     <script type="module" src="/src/layouts/Header.js"></script>
   </head>
   <body>
     <div class="container">
       <!-- 프로필 헤더 -->
       <div class="ProfileHeader">
-        <profile-photo></profile-photo>
+        <profile-header></profile-header>
+        <profile-icon></profile-icon>
       </div>
 
       <!-- 매너 온도 및 리뷰 섹션 -->
